+++ backend/de.metas.business/src/main/java/de/metas/order/IOrderBL.java
@@ -283,4 +283,8 @@ public interface IOrderBL extends ISingletonService
 
 	void updateIsOnConsignmentFromLines(OrderId orderId);
 
+	void markAsDropShipIfDropShipWarehouse(I_C_Order order);
+
+	void setBillToDefaultLocationAsDefaultIfDropShipSO(I_C_Order order);
+
 }
